Build i686 runtimes for the host
Bug: http://b/28196329

These are used for 32-bit asan targets on the host.

Due to CMake and Clang oddities, we need to explicitly set
CMAKE_C_COMPILER_TARGET and pass '-march=i686' in the compiler flags.
This is because the default CPU for i686 is pentium4, which skips
definition of the __i686__ macro.  But the compiler-rt CMake flags rely
on the presence __i686__ macro to add i686 as one of the supported
archs.

Test: Build toolchain and use it to build the platform with
      'SANITIZE_HOST=address'

Change-Id: Id4c171833acbb380fa1bb6221fd06275e74f61d3