Disable SANITIZER_CAN_USE_CXXABI for stage2 build
Bug: http://b/28196329

This causes the libclang_rt.asan_cxx*.a for the host to depend on
libc++abi.  We don't want that since there are platform modules that
link without libc++/libc++abi and will fail to build with asan.

An ideal fix will be to add -fno-sanitizer-link-c++-runtime (to
complement -fsanitizer-link-c++-runtime) and pass this to 'stl: none'
modules on host-sanitizer builds.

Test: Use prebuilts to build Android with SANITIZE_HOST=address
Change-Id: Ic46dd3c3c08940256f296ff58d1e08d4c0f17560