Handle the case where some prebuilts are already generated.
Bug: http://b/68397894
Test: ./update-prebuilts.py --skip-fetch --skip-cleanup -b 68397894 4556391

If we already have prebuilts in some directories (but possibly not
others), it makes sense to just skip the empty commit. The empty commit
will fail the check_call(), when it really is mostly harmless. Nothing
would get uploaded later because of the empty commit, and that should be
enough of an indicator to the developer that something out of the
ordinary is happening. I have already used this script to update just
the darwin prebuilts for the case of build 4556391. In this case, the
prior upload was only broken for darwin, so running this script allowed
me to generate just the darwin prebuilts, while skipping linux/windows
as I expected.

Change-Id: I0d56e751268e990c44543320f19b83e99bc0b1b4