Add svn_revision information to our version information.
This adds svn_revision to android_version.py, so that we can use it in
our CLANG_VENDOR string, as well as putting it into AndroidVersion.txt.
Even though this svn information is usually misleading (due to many
cherry-picks), it is probably easier for some people to reason about
Clang features that are present/missing.

Bug: http://b/68941234
Test: ./build.py and check clang --version and AndroidVersion.txt.
Change-Id: I05cbbf51024de2b9a760c05c5db96ef22f5aea31