Enable openmp static library build for stage2
Bug: http://b/64121881

Disabling host OpenMP build
(https://android-review.googlesource.com/#/c/625229/) also omitted the
installation of omp.h, which is needed for OpenMP on the device.  Build
the static OpenMP library instead.  The omp.h header is included, but
the built lib*omp*.a are discarded during packaging (along with all
static libraries).  This is also cleaner because we cannot use these for
OpenMP on the host (due to mismatched libc++).

Test: build.py and check lib64/clang/6.0.1/include/omp.h is present.
Change-Id: I389068fe7c2c4328116af1cf5efd7d3420efacae