Add update-lldb-prebuilts.py script
The script downloads lldb binaries from the build servers and extracts
them to the prebuilts folder (clang/host/linux-x86/lldb), ready for
check-in. I tried to preserve the interface of update-prebuilts.py as
much as possible.

Only linux host is currently supported, and I still need to add utility
scripts which would make it usable for debugging android applications.

Bug: 76097225
Change-Id: I9ee67a831f7f37b7eba6f622d94e0e0f805ff05d