Strip added flags when calling the fallback compiler
The compiler wrapper adds some flags (primarily to silence warnings)
when testing the new compiler (via test_compiler.py).  These warnings
must be stripped when falling back to the old compiler else the fallback
will not work as intended.

This is causing a build break now since the fallback does not recognize
-fsplit-lto-unit.

Test: Build x86_64 with test_compiler.py
Change-Id: I368596475855c7daccbf2fdb89b57dc0b18ad285