Build libc++abi and use it to build compiler-rt
Bug: http://b/119284374
Bug: http://b/126457671
Bug: http://b/128937182

Compiler-rt runtimes like asan, hwasan need libc++abi to get typeinfo of
objects at runtime.  Currently a stub libc++abi library, redirected to a
stub libc++.so, is used when building these runtimes.  The libraries end
up depending on libc++.so.  At execution time, they'll use the libc++
from the platform.  This is not ideal since it's interesting to sanitize
the libc++ in the platform as well.

This change builds a libc++abi from toolchain/libcxxabi and uses it when
building compiler-rt runtimes.  This change is restricted to arm64 and
x86_64 since this is mainly blocking hwasan (which is 64-bit only) right
now.

Test: Verify that compiler-rt runtimes (asan, hwasan, scudo, etc.) for
aarch64, x86_64 do not link against libc++.so

Change-Id: Icbf74085a23d0d45ef14fa073c8644e6b7d4c882