Pass __ANDROID_API__=10000 when building platform libraries
Bug; http://b/138304776

The platform variant of the compiler-rt runtimes are built using the NDK
with an API level of 29.  We should be building with __ANDORID_API__ set
to __ANDROID_API_FUTURE__ and linking against libc, etc. built from the
platform.

This change is needed now for upstream change
https://reviews.llvm.org/D65245 that checks __ANDROID_API__ to do
something custom for the platform.

Test: build.py and master-plus-llvm
Change-Id: I3c04c8c59781d5d031fe6ce41505131d5189d9ca