[kythe] Do not build libcxx benchmarks
Bug: http://b/151096875

CMAKE_MAKE_PROGRAM is not passed to the CMake sub-build that builds
libcxx for the benchmark.  This causes the benchmark build to fail
on the build server (which does not have a Ninja pre-installed).

Also change how expected failures are checked - ignore failures in any
source in the scudo and libcxx benchmark dirs.

Test: kythe_xref.py
Change-Id: I0caa26289f1708bbb9101bb834a053b13b1b52bd