Do not set patch_level for llvm-next builds
The patch level is only relevant for the release toolchains.  Using the
same patch level for both release and llvm-next toolchains may cause the
version of the llvm-next toolchains to be non-linear.

In addition, there are  few other changes:
- Do not set CLANG_VERSION_PATCHLEVEL.  If unset, the build system will
initialize it to LLVM_VERSION_PATCH.
- Refactor android_version.py to remove unnecessary uses of llvm_next
parameters.  That flag is going to be fixed during an invocation of this
script.

Change-Id: If3dcc528c16f79fb3d7c1492c3e2f675161492ca