[test scripts] Record CL information to CNS
The CNS directory has two CSV files now:
- prebuilt_cls.csv: with info about clang prebuilt CLs
- soong_cls.csv: with info about switchover CLs in build/soong.

test_prebuilts.py takes 3 arguments now:
  --build <build> - a build number in go/ab/.
  --prebuilt_cl <cl> - (optional) gerrit CL uploading new prebuilts.
  --soong_cl <cl> - (optional) gerrit CL switching clang version in build/soong

On success, the script would have found up to two CLs to test together
using Forrest (one if the prebuilts are already merged).  If no CLs are
found in the CSV files and none were provided in the command line, the
script will create/update CLs to gerrit.  The script will also validate
that the scripts are consistent and testable (e.g. no merge conflicts).

(NB: The output is a bit verbose and can be cleaned up in follow-ups.)

Test: test_prebuilts.py 6842337 created changes 1429142, 1429608 and
updated records in the CNS directory.

Change-Id: I3205d0b10591ce650904a940a6c60fff54526dc2