[test scripts] Try to resolve merge conflicts in soong CL
If a previously recorded gerrit CL has a merge conflict, and if that CL
only updates clang version strings, try to resolve the conflicts.  Do so
by create a new commit but with the Change-ID of the old CL.

Also always call `repo sync` in build/soong before creating CLs.

Test: test_prebuilts.py --build 6842337 --prepare-only resolved
      conflicts in aosp/1429608
Change-Id: Ie1e2a494d284ef0017bf2874af3cf48363c8895d