Use logger.debug() in utils.subprocess_run
... and drop down the logging level of scripts to DEBUG.  This preserves
the current output levels of these scripts.

Also remove the `--verbose` option to build.py.  It's a no-op (both
before and after this change).  We can add it back if we want to reduce
the verbosity of build.py.

The motivation behind this change is to reduce verbose output from
test/scripts/* - have the DEBUG level to print the verbose output and
the INFO level print progress at a coarser level ( `Uploaded CL to
build/soong` or `Invoked forrest run <>` etc.)

Test: build.py and check logging.
Change-Id: I548a875f6911ba4fab3647c137898b9b06aca834