Turn LIBUNWIND_ENABLE_ASSERTIONS off by default
Set LIBUNWIND_ENABLE_ASSERTIONS according to self.enable_assertions.

In the CMake file, LIBUNWIND_ENABLE_ASSERTIONS defaults to 'on'.
Leaving it on might slow unwinding, and it results in three debugging
functions output into libunwind*.a: logAPIs, logDWARF, logUnwinding.
We should omit these functions because the names aren't reserved.

With this change, the libunwind*.a prebuilts are compiled with
-DNDEBUG, like libunwind_llvm.a in the platform.

Bug: http://b/153025717
Test: "nm out-llvm/lib/libunwind-*/lib64/libunwind.a | grep logAPIs"
Change-Id: I548d8ddad0f3078938dbe99c7962d73f430a39a1