Revert "[kythe] Use SHA of toolchain/llvm-project in output"
This reverts commit c03774b4b089e84224d9f46c4c88349261daad6a.

Reason for revert: publish artifacts can now support <SHA.kzip>

Original commit message:
The SHA will be used by the kythe backend to match crossreferences to
the source.  Currently, the build will fail if the SHA is not available
on the build server.  We can add a fallback or alternative way of
getting the SHA if the build failures are too noisy.

Bug: http://b162557156
Change-Id: Ic0a3ddd627f7f4692b2991e743153e932d2326ff
Test: toolchain/llvm_android/kythe_xref.py