Adapt cherrypick_cl.py for our latest JSON schema.
"Now that we've altered the schema for PATCHES.json, I'm altering
cherrypick_cl.py. Pray I don't alter it any further."

Our upstream cherrypick script needed to be updated to reflect the
latest schema for PATCHES.json. This change reorders the fields of
PatchItem to reflect the sorted order (which we also dump explicitly as
sorted now). It also adds a newline to preserve the current state of the
file for no-op runs of the script.

Test: ./cherrypick_cl.py --reason Testing --sha 6774cc33f75f8c0d6a6e1e41b6935f243e581710  # adds one new patch properly
Test: ./cherrypick_cl.py --reason Testing  # is identical
Change-Id: Ic73741ea02a682542dd622c2b43c1f2b5ade3c8f