[patches] Cherry pick CLS for: ARCH=arm Linux kernel builds
da66263b6e5 [ARM] implement support for ALU/LDR PC-relative group relocations

The Linux kernel will make use of these relocations in a patch set that
is slated to land in 5.17-rc1 in a few weeks:

https://lore.kernel.org/r/20211122092816.2865873-8-ardb@kernel.org/

The patch set has guarded the use of these relocations on LLVM versions
14.0.0 or newer, as that is the released version that will first support
them. The kernel's policy around LLVM version checks has always assumed
that groups using or basing on unreleased LLVM versions are updating
them frequently enough to avoid breakage due to the toolchain version
claiming to have support for the feature without really having support
for the feature.

In Android's case, the toolchain version is 14.0.0 but it does not
have support for these relocations so the build fails:

https://builds.tuxbuild.com/23K7btd5Hed4vANSvRmQbd1Q94X/build.log

Add the patch that adds support for these relocations so there is no
more build breakage.

Test: Broken Linux kernels now build and boot successfully
Change-Id: Ieb5f648bfa051a6bc38808a60a23205e7ca0c04a
Signed-off-by: Nathan Chancellor <nathan@kernel.org>