Fix LLVM runtimes compiler
Bug: http://b/250198090
Bug: http://b/256191783

Use LLVM_ENABLE_RUNTIMES to build compiler-rt;libcxx;libcxxabi for the
host instead of LLVM_ENABLED_PROJECTS.  The latter is deprecated and
also causes issues with the Darwin build.

For musl, we need to install library dependencies in two steps - to
out/stage{1,2} before build so we can use the just-built clang and lld
to build the runtimes, and to out/stage{1,2}-install after the
install.

Test: build.py on linux, darwin, musl

Change-Id: Iafa465021e8616ba2c4e0db7a0ef4f60acc94cba