Stop versioning LLVM libs (libc++, libc++abi, libclang-cpp)
Generate libc++[abi].{so,dylib} without version numbers in the filename
or the DT_SONAME. This is needed for Soong, which copies the host
libc++ to a path without a version number, making it inaccessible if
the files using it have a DT_NEEDED (or LC_LOAD_DYLIB) with the version
number.

For consistency with Linux, include libc++abi.a's objects in Darwin's
libc++.{a,dylib} libraries.

For backwards compatibility, generate symlinks:
 - libc++[abi].so.1 to libc++[abi].so
 - libc++[abi].1.dylib to libc++[abi].dylib

Bug: http://b/175635923
Test: treehugger
Change-Id: Ib7b3af998239937b92e9089af4ffda0c74c3a324