Add LibSimpleperfReadElfBuilder
It is used to build prebuilt static llvm libraries used by simpleperf.
For linux and darwin hosts, the libraries are collected from stage2
build. For devices and windows host, the libraries are built by
LibSimpleperfReadElfBuilder.

To make the libraries easier to be used in the Android build system,
archive them into one file called libsimpleperf_readelf.a.

Bug: 267670561
Test: run build.py on linux/darwin.
Change-Id: I83d135cbe279c082ba8d79142237509ac66c999d