Use the sysroot's libc++ for Android platform LLVM builds
This change is necessary for building the fuzzer library with a custom
libc++ (COMPILER_RT_USE_LIBCXX), because the fuzzer needs to add its
own libc++ headers to the search path, and it's not possible to have
two sets of libc++ headers at once.

e.g. The stddef.h in the first libc++ will delegate to the stddef.h in
the second, which will see that `_LIBCPP_STDDEF_H` is already defined,
and then not delegate to the stddef.h in the clang resource directory.
As a result, size_t/ptrdiff_t/etc aren't defined.

Bug: b/175635923
Bug: b/303175229
Test: build.py
Change-Id: I1edfc53dc7000d360e6a9c4fa4d3288f2f1fd4a7