Build a custom libc++ into the fuzzer library
For arm64, the fuzzer library is not built with HWASan, but object
files in the application might emit some libc++ symbols with HWASan
enabled, like `std::__1::piecewise_construct`. Link an isolated copy
of libc++ into the fuzzer library to avoid HWASan<->non-HWASan
incompatibility.

The fuzzer LLVM patch is currently being reviewed in:
https://github.com/llvm/llvm-project/pull/70407

Bug: b/175635923
Bug: b/303175229
Test: build.py
Change-Id: Ie6b8761c2298bd168b4389c45df9698f8046276d